refactor: extract DeviceMonitor component to eliminate code duplication
- Create DeviceMonitor component to centralize video stream monitoring logic
- Remove ~200 lines of duplicated code from DevicePanel and ChatKitPanel
- Fix mode switching issue: keep both panels mounted with CSS hidden instead of unmounting
- Video stream no longer resets when switching between classic/layered agent modes

This follows DRY principle and improves code maintainability.
