+++ frontend/src/components/ChatKitPanel.tsx
@@ -5,15 +5,8 @@ import { Textarea } from '@/components/ui/textarea';
 import { Badge } from '@/components/ui/badge';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { useTranslation } from '../lib/i18n-context';
-import { ScrcpyPlayer } from './ScrcpyPlayer';
+import { DeviceMonitor } from './DeviceMonitor';
 import {
-  Video,
-  Image as ImageIcon,
-  MonitorPlay,
-  ChevronLeft,
-  ChevronRight,
-  Fingerprint,
-  ArrowUpDown,
   AlertCircle,
   CheckCircle2,
   Loader2,
@@ -28,8 +21,8 @@ import {
   ListChecks,
   Square,
 } from 'lucide-react';
-import type { ScreenshotResponse, Workflow } from '../api';
-import { getScreenshot, listWorkflows, getErrorMessage } from '../api';
+import type { Workflow } from '../api';
+import { listWorkflows, getErrorMessage } from '../api';
 import {
   Popover,
   PopoverContent,
@@ -98,30 +91,6 @@ export function ChatKitPanel({
   const eventSourceRef = React.useRef<EventSource | null>(null);
   const abortControllerRef = React.useRef<AbortController | null>(null);
 
-  // Screen display state
-  const [useVideoStream, setUseVideoStream] = React.useState(true);
-  const [videoStreamFailed, setVideoStreamFailed] = React.useState(false);
-  const [displayMode, setDisplayMode] = React.useState<
-    'auto' | 'video' | 'screenshot'
-  >('auto');
-  const [screenshot, setScreenshot] = React.useState<ScreenshotResponse | null>(
-    null
-  );
-  const [feedbackMessage, setFeedbackMessage] = React.useState<string | null>(
-    null
-  );
-  const [feedbackType, setFeedbackType] = React.useState<
-    'tap' | 'swipe' | 'error' | 'success'
-  >('success');
-  const feedbackTimeoutRef = React.useRef<number | null>(null);
-
-  // Control area state
-  const [showControlArea, setShowControlArea] = React.useState(false);
-  const [showControls, setShowControls] = React.useState(false);
-  const controlsTimeoutRef = React.useRef<number | null>(null);
-  const videoStreamRef = React.useRef<{ close: () => void } | null>(null);
-  const screenshotFetchingRef = React.useRef(false);
-
   // Workflow state
   const [workflows, setWorkflows] = React.useState<Workflow[]>([]);
   const [showWorkflowPopover, setShowWorkflowPopover] = React.useState(false);
@@ -130,70 +99,17 @@ export function ChatKitPanel({
   const [historyItems, setHistoryItems] = React.useState<HistoryItem[]>([]);
   const [showHistoryPopover, setShowHistoryPopover] = React.useState(false);
 
-  const showFeedback = (
-    message: string,
-    duration = 2000,
-    type: 'tap' | 'swipe' | 'error' | 'success' = 'success'
-  ) => {
-    if (feedbackTimeoutRef.current) {
-      clearTimeout(feedbackTimeoutRef.current);
-    }
-    setFeedbackType(type);
-    setFeedbackMessage(message);
-    feedbackTimeoutRef.current = setTimeout(() => {
-      setFeedbackMessage(null);
-    }, duration);
-  };
-
-  // Scroll to bottom when messages change
   React.useEffect(() => {
     messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
   }, [messages]);
 
-  // Cleanup on unmount
   React.useEffect(() => {
     return () => {
-      if (feedbackTimeoutRef.current) {
-        clearTimeout(feedbackTimeoutRef.current);
-      }
       if (eventSourceRef.current) {
         eventSourceRef.current.close();
       }
     };
   }, []);
-
-  const handleMouseEnter = () => {
-    if (controlsTimeoutRef.current) {
-      clearTimeout(controlsTimeoutRef.current);
-    }
-    setShowControlArea(true);
-  };
-
-  const handleMouseLeave = () => {
-    controlsTimeoutRef.current = setTimeout(() => {
-      setShowControlArea(false);
-    }, 500);
-  };
-
-  React.useEffect(() => {
-    return () => {
-      if (controlsTimeoutRef.current) {
-        clearTimeout(controlsTimeoutRef.current);
-      }
-    };
-  }, []);
-
-  const toggleControls = () => {
-    setShowControls(prev => !prev);
-  };
-
-  React.useEffect(() => {
-    return () => {
-      if (videoStreamRef.current) {
-        videoStreamRef.current.close();
-      }
-    };
-  }, [deviceId]);
 
   // Load workflows
   React.useEffect(() => {
@@ -251,56 +167,6 @@ export function ChatKitPanel({
   const handleDeleteHistoryItem = (itemId: string) => {
     deleteHistoryItem(deviceSerial, itemId);
     setHistoryItems(prev => prev.filter(item => item.id !== itemId));
-  };
-
-  // Screenshot polling
-  React.useEffect(() => {
-    if (!deviceId || !isVisible) return;
-
-    const shouldPollScreenshots =
-      displayMode === 'screenshot' ||
-      (displayMode === 'auto' && videoStreamFailed);
-
-    if (!shouldPollScreenshots) {
-      return;
-    }
-
-    const fetchScreenshot = async () => {
-      if (screenshotFetchingRef.current) return;
-
-      screenshotFetchingRef.current = true;
-      try {
-        const data = await getScreenshot(deviceId);
-        if (data.success) {
-          setScreenshot(data);
-        }
-      } catch (e) {
-        console.error('Failed to fetch screenshot:', e);
-      } finally {
-        screenshotFetchingRef.current = false;
-      }
-    };
-
-    fetchScreenshot();
-    const interval = setInterval(fetchScreenshot, 500);
-
-    return () => clearInterval(interval);
-  }, [deviceId, videoStreamFailed, displayMode, isVisible]);
-
-  const handleVideoStreamReady = React.useCallback(
-    (stream: { close: () => void } | null) => {
-      videoStreamRef.current = stream;
-    },
-    []
-  );
-
-  const handleFallback = React.useCallback(() => {
-    setVideoStreamFailed(true);
-    setUseVideoStream(false);
-  }, []);
-
-  const toggleDisplayMode = (mode: 'auto' | 'video' | 'screenshot') => {
-    setDisplayMode(mode);
   };
 
   // Toggle step expansion
@@ -613,12 +479,8 @@ export function ChatKitPanel({
           console.error('Backend abort failed:', e)
         );
       });
-
-      // Show feedback
-      showFeedback(t.chat?.aborted || '任务已中断', 2000, 'success');
     } catch (error) {
       console.error('Failed to abort chat:', error);
-      showFeedback(t.chat?.abortFailed || '中断失败', 2000, 'error');
     } finally {
       setLoading(false);
       setAborting(false);
@@ -1065,195 +927,7 @@ export function ChatKitPanel({
         </div>
       </Card>
 
-      {/* Screen preview - phone aspect ratio */}
-      <Card
-        className="w-[320px] flex-shrink-0 relative min-h-0 overflow-hidden bg-background"
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-      >
-        {/* Toggle and controls - shown on hover */}
-        <div
-          className={`absolute top-4 right-4 z-10 transition-opacity duration-200 ${
-            showControlArea ? 'opacity-100' : 'opacity-0 pointer-events-none'
-          }`}
-        >
-          <div className="flex items-center gap-2">
-            {/* Control buttons - slide in/out */}
-            <div
-              className={`flex items-center gap-1 bg-popover/90 backdrop-blur rounded-xl p-1 shadow-lg border border-border transition-all duration-300 ${
-                showControls
-                  ? 'opacity-100 translate-x-0'
-                  : 'opacity-0 translate-x-4 pointer-events-none'
-              }`}
-            >
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('auto')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'auto'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                {t.devicePanel?.auto || 'Auto'}
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('video')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'video'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                <Video className="w-3 h-3 mr-1" />
-                {t.devicePanel?.video || 'Video'}
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('screenshot')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'screenshot'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                <ImageIcon className="w-3 h-3 mr-1" />
-                {t.devicePanel?.image || 'Image'}
-              </Button>
-            </div>
-
-            {/* Toggle button - visible when control area is shown */}
-            <Button
-              variant="ghost"
-              size="icon"
-              onClick={toggleControls}
-              className="h-8 w-8 rounded-full bg-popover/90 backdrop-blur border border-border shadow-lg hover:bg-accent"
-              title={showControls ? 'Hide controls' : 'Show controls'}
-            >
-              {showControls ? (
-                <ChevronRight className="w-4 h-4" />
-              ) : (
-                <ChevronLeft className="w-4 h-4" />
-              )}
-            </Button>
-          </div>
-        </div>
-
-        {/* Current mode indicator - bottom left */}
-        <div className="absolute bottom-4 left-4 z-10">
-          <Badge
-            variant="secondary"
-            className="bg-white/90 text-slate-700 border border-slate-200 dark:bg-slate-900/90 dark:text-slate-300 dark:border-slate-700"
-          >
-            {displayMode === 'auto' && (t.devicePanel?.auto || 'Auto')}
-            {displayMode === 'video' && (
-              <>
-                <MonitorPlay className="w-3 h-3 mr-1" />
-                {t.devicePanel?.video || 'Video'}
-              </>
-            )}
-            {displayMode === 'screenshot' && (
-              <>
-                <ImageIcon className="w-3 h-3 mr-1" />
-                {t.devicePanel?.imageRefresh || 'Screenshot'}
-              </>
-            )}
-          </Badge>
-        </div>
-
-        {/* Feedback message */}
-        {feedbackMessage && (
-          <div className="absolute bottom-4 right-4 z-20 flex items-center gap-2 px-3 py-2 bg-[#1d9bf0] text-white text-sm rounded-xl shadow-lg">
-            {feedbackType === 'error' && <AlertCircle className="w-4 h-4" />}
-            {feedbackType === 'tap' && <Fingerprint className="w-4 h-4" />}
-            {feedbackType === 'swipe' && <ArrowUpDown className="w-4 h-4" />}
-            {feedbackType === 'success' && <CheckCircle2 className="w-4 h-4" />}
-            <span>{feedbackMessage}</span>
-          </div>
-        )}
-
-        {/* Video stream */}
-        {displayMode === 'video' ||
-        (displayMode === 'auto' && useVideoStream && !videoStreamFailed) ? (
-          <ScrcpyPlayer
-            deviceId={deviceId}
-            className="w-full h-full"
-            enableControl={true}
-            onFallback={handleFallback}
-            onTapSuccess={() =>
-              showFeedback(t.devicePanel?.tapped || 'Tapped', 2000, 'tap')
-            }
-            onTapError={error =>
-              showFeedback(
-                (t.devicePanel?.tapError || 'Tap error: {error}').replace(
-                  '{error}',
-                  error
-                ),
-                3000,
-                'error'
-              )
-            }
-            onSwipeSuccess={() =>
-              showFeedback(t.devicePanel?.swiped || 'Swiped', 2000, 'swipe')
-            }
-            onSwipeError={error =>
-              showFeedback(
-                (t.devicePanel?.swipeError || 'Swipe error: {error}').replace(
-                  '{error}',
-                  error
-                ),
-                3000,
-                'error'
-              )
-            }
-            onStreamReady={handleVideoStreamReady}
-            fallbackTimeout={100000}
-          />
-        ) : (
-          /* Screenshot mode */
-          <div className="w-full h-full flex items-center justify-center bg-muted/30 min-h-0">
-            {screenshot && screenshot.success ? (
-              <div className="relative w-full h-full flex items-center justify-center min-h-0">
-                <img
-                  src={`data:image/png;base64,${screenshot.image}`}
-                  alt="Device Screenshot"
-                  className="max-w-full max-h-full object-contain"
-                  style={{
-                    width:
-                      screenshot.width > screenshot.height ? '100%' : 'auto',
-                    height:
-                      screenshot.width > screenshot.height ? 'auto' : '100%',
-                  }}
-                />
-                {screenshot.is_sensitive && (
-                  <div className="absolute top-12 right-2 px-2 py-1 bg-yellow-500 text-white text-xs rounded-lg">
-                    {t.devicePanel?.sensitiveContent || 'Sensitive Content'}
-                  </div>
-                )}
-              </div>
-            ) : screenshot?.error ? (
-              <div className="text-center text-destructive">
-                <AlertCircle className="w-8 h-8 mx-auto mb-2" />
-                <p className="font-medium">
-                  {t.devicePanel?.screenshotFailed || 'Screenshot Failed'}
-                </p>
-                <p className="text-xs mt-1 opacity-60">{screenshot.error}</p>
-              </div>
-            ) : (
-              <div className="text-center text-muted-foreground">
-                <Loader2 className="w-8 h-8 mx-auto mb-2 animate-spin" />
-                <p className="text-sm">
-                  {t.devicePanel?.loading || 'Loading...'}
-                </p>
-              </div>
-            )}
-          </div>
-        )}
-      </Card>
+      <DeviceMonitor deviceId={deviceId} isVisible={isVisible} />
     </div>
   );
 }
