+++ frontend/src/components/DevicePanel.tsx
@@ -6,13 +6,6 @@ import {
   AlertCircle,
   Loader2,
   Sparkles,
-  Video,
-  Image as ImageIcon,
-  MonitorPlay,
-  Fingerprint,
-  ArrowUpDown,
-  ChevronLeft,
-  ChevronRight,
   History,
   ListChecks,
   Square,
@@ -21,11 +14,10 @@ import {
   Rocket,
 } from 'lucide-react';
 import { throttle } from 'lodash';
-import { ScrcpyPlayer } from './ScrcpyPlayer';
+import { DeviceMonitor } from './DeviceMonitor';
 import { DualModelPanel } from './DualModelPanel';
 import { useDualModelState } from './useDualModelState';
 import type {
-  ScreenshotResponse,
   ThinkingChunkEvent,
   StepEvent,
   DoneEvent,
@@ -35,7 +27,6 @@ import type {
 } from '../api';
 import {
   abortChat,
-  getScreenshot,
   initAgent,
   resetChat,
   sendMessageStream,
@@ -127,16 +118,6 @@ export function DevicePanel({
   const [aborting, setAborting] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [initialized, setInitialized] = useState(false);
-  const [screenshot, setScreenshot] = useState<ScreenshotResponse | null>(null);
-  const [useVideoStream, setUseVideoStream] = useState(true);
-  const [videoStreamFailed, setVideoStreamFailed] = useState(false);
-  const [displayMode, setDisplayMode] = useState<
-    'auto' | 'video' | 'screenshot'
-  >('auto');
-  const [feedbackMessage, setFeedbackMessage] = useState<string | null>(null);
-  const [feedbackType, setFeedbackType] = useState<
-    'tap' | 'swipe' | 'error' | 'success'
-  >('success');
   const [showHistoryPopover, setShowHistoryPopover] = useState(false);
   const [historyItems, setHistoryItems] = useState<HistoryItem[]>([]);
   const [workflows, setWorkflows] = useState<Workflow[]>([]);
@@ -147,66 +128,11 @@ export function DevicePanel({
     handleEvent: handleDualModelEvent,
     reset: resetDualModelState,
   } = useDualModelState();
-  const feedbackTimeoutRef = useRef<number | null>(null);
-
-  const showFeedback = (
-    message: string,
-    duration = 2000,
-    type: 'tap' | 'swipe' | 'error' | 'success' = 'success'
-  ) => {
-    if (feedbackTimeoutRef.current) {
-      clearTimeout(feedbackTimeoutRef.current);
-    }
-    setFeedbackType(type);
-    setFeedbackMessage(message);
-    feedbackTimeoutRef.current = setTimeout(() => {
-      setFeedbackMessage(null);
-    }, duration);
-  };
-
-  useEffect(() => {
-    return () => {
-      if (feedbackTimeoutRef.current) {
-        clearTimeout(feedbackTimeoutRef.current);
-      }
-    };
-  }, []);
-
-  const [showControlArea, setShowControlArea] = useState(false);
-  const [showControls, setShowControls] = useState(false);
-  const controlsTimeoutRef = useRef<number | null>(null);
-
-  const handleMouseEnter = () => {
-    if (controlsTimeoutRef.current) {
-      clearTimeout(controlsTimeoutRef.current);
-    }
-    setShowControlArea(true);
-  };
-
-  const handleMouseLeave = () => {
-    controlsTimeoutRef.current = setTimeout(() => {
-      setShowControlArea(false);
-    }, 500);
-  };
-
-  useEffect(() => {
-    return () => {
-      if (controlsTimeoutRef.current) {
-        clearTimeout(controlsTimeoutRef.current);
-      }
-    };
-  }, []);
-
-  const toggleControls = () => {
-    setShowControls(prev => !prev);
-  };
 
   const chatStreamRef = useRef<{ close: () => void } | null>(null);
   const dualModelStreamRef = useRef<{ close: () => void } | null>(null);
-  const videoStreamRef = useRef<{ close: () => void } | null>(null);
   const messagesEndRef = useRef<HTMLDivElement>(null);
   const messagesContainerRef = useRef<HTMLDivElement>(null);
-  const screenshotFetchingRef = useRef(false);
   const hasAutoInited = useRef(false);
   const prevConfigRef = useRef<GlobalConfig | null>(null);
   const prevMessageCountRef = useRef(0);
@@ -628,11 +554,6 @@ export function DevicePanel({
         );
         setLoading(false);
         chatStreamRef.current = null;
-
-        // Show feedback
-        setFeedbackMessage(t.chat.aborted);
-        setFeedbackType('success');
-        setTimeout(() => setFeedbackMessage(null), 2000);
       }
     );
 
@@ -645,9 +566,6 @@ export function DevicePanel({
     deviceSerial,
     deviceName,
     handleInit,
-    t,
-    setFeedbackMessage,
-    setFeedbackType,
   ]);
 
   // Dual model send function
@@ -861,16 +779,8 @@ export function DevicePanel({
           console.error('Backend abort failed:', e)
         );
       }
-
-      // Show feedback
-      setFeedbackMessage(t.chat.aborted);
-      setFeedbackType('success');
-      setTimeout(() => setFeedbackMessage(null), 2000);
     } catch (error) {
       console.error('Failed to abort chat:', error);
-      setFeedbackMessage(t.chat.abortFailed);
-      setFeedbackType('error');
-      setTimeout(() => setFeedbackMessage(null), 2000);
     } finally {
       setLoading(false);
       setAborting(false);
@@ -928,9 +838,6 @@ export function DevicePanel({
       if (dualModelStreamRef.current) {
         dualModelStreamRef.current.close();
       }
-      if (videoStreamRef.current) {
-        videoStreamRef.current.close();
-      }
     };
   }, [deviceId]);
 
@@ -964,39 +871,6 @@ export function DevicePanel({
     setIsAtBottom(true);
   };
 
-  useEffect(() => {
-    if (!deviceId) return;
-
-    const shouldPollScreenshots =
-      displayMode === 'screenshot' ||
-      (displayMode === 'auto' && videoStreamFailed);
-
-    if (!shouldPollScreenshots) {
-      return;
-    }
-
-    const fetchScreenshot = async () => {
-      if (screenshotFetchingRef.current) return;
-
-      screenshotFetchingRef.current = true;
-      try {
-        const data = await getScreenshot(deviceId);
-        if (data.success) {
-          setScreenshot(data);
-        }
-      } catch (e) {
-        console.error('Failed to fetch screenshot:', e);
-      } finally {
-        screenshotFetchingRef.current = false;
-      }
-    };
-
-    fetchScreenshot();
-    const interval = setInterval(fetchScreenshot, 500);
-
-    return () => clearInterval(interval);
-  }, [deviceId, videoStreamFailed, displayMode]);
-
   const handleInputKeyDown = (
     event: React.KeyboardEvent<HTMLTextAreaElement>
   ) => {
@@ -1004,22 +878,6 @@ export function DevicePanel({
       event.preventDefault();
       handleSendMessage();
     }
-  };
-
-  const handleVideoStreamReady = useCallback(
-    (stream: { close: () => void } | null) => {
-      videoStreamRef.current = stream;
-    },
-    []
-  );
-
-  const handleFallback = useCallback(() => {
-    setVideoStreamFailed(true);
-    setUseVideoStream(false);
-  }, []);
-
-  const toggleDisplayMode = (mode: 'auto' | 'video' | 'screenshot') => {
-    setDisplayMode(mode);
   };
 
   return (
@@ -1525,183 +1383,7 @@ export function DevicePanel({
         </div>
       </Card>
 
-      {/* Screen preview - phone aspect ratio */}
-      <Card
-        className="w-[320px] flex-shrink-0 relative min-h-0 overflow-hidden bg-background"
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-      >
-        {/* Toggle and controls - shown on hover */}
-        <div
-          className={`absolute top-4 right-4 z-10 transition-opacity duration-200 ${
-            showControlArea ? 'opacity-100' : 'opacity-0 pointer-events-none'
-          }`}
-        >
-          <div className="flex items-center gap-2">
-            {/* Control buttons - slide in/out */}
-            <div
-              className={`flex items-center gap-1 bg-popover/90 backdrop-blur rounded-xl p-1 shadow-lg border border-border transition-all duration-300 ${
-                showControls
-                  ? 'opacity-100 translate-x-0'
-                  : 'opacity-0 translate-x-4 pointer-events-none'
-              }`}
-            >
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('auto')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'auto'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                {t.devicePanel.auto}
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('video')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'video'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                <Video className="w-3 h-3 mr-1" />
-                {t.devicePanel.video}
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => toggleDisplayMode('screenshot')}
-                className={`h-7 px-3 text-xs rounded-lg transition-colors ${
-                  displayMode === 'screenshot'
-                    ? 'bg-primary text-primary-foreground'
-                    : 'text-foreground hover:bg-accent hover:text-accent-foreground'
-                }`}
-              >
-                <ImageIcon className="w-3 h-3 mr-1" />
-                {t.devicePanel.image}
-              </Button>
-            </div>
-
-            {/* Toggle button - visible when control area is shown */}
-            <Button
-              variant="ghost"
-              size="icon"
-              onClick={toggleControls}
-              className="h-8 w-8 rounded-full bg-popover/90 backdrop-blur border border-border shadow-lg hover:bg-accent"
-              title={showControls ? 'Hide controls' : 'Show controls'}
-            >
-              {showControls ? (
-                <ChevronRight className="w-4 h-4" />
-              ) : (
-                <ChevronLeft className="w-4 h-4" />
-              )}
-            </Button>
-          </div>
-        </div>
-
-        {/* Current mode indicator - bottom left */}
-        <div className="absolute bottom-4 left-4 z-10">
-          <Badge
-            variant="secondary"
-            className="bg-white/90 text-slate-700 border border-slate-200 dark:bg-slate-900/90 dark:text-slate-300 dark:border-slate-700"
-          >
-            {displayMode === 'auto' && t.devicePanel.auto}
-            {displayMode === 'video' && (
-              <>
-                <MonitorPlay className="w-3 h-3 mr-1" />
-                {t.devicePanel.video}
-              </>
-            )}
-            {displayMode === 'screenshot' && (
-              <>
-                <ImageIcon className="w-3 h-3 mr-1" />
-                {t.devicePanel.imageRefresh}
-              </>
-            )}
-          </Badge>
-        </div>
-
-        {/* Feedback message */}
-        {feedbackMessage && (
-          <div className="absolute bottom-4 right-4 z-20 flex items-center gap-2 px-3 py-2 bg-[#1d9bf0] text-white text-sm rounded-xl shadow-lg">
-            {feedbackType === 'error' && <AlertCircle className="w-4 h-4" />}
-            {feedbackType === 'tap' && <Fingerprint className="w-4 h-4" />}
-            {feedbackType === 'swipe' && <ArrowUpDown className="w-4 h-4" />}
-            {feedbackType === 'success' && <CheckCircle2 className="w-4 h-4" />}
-            <span>{feedbackMessage}</span>
-          </div>
-        )}
-
-        {/* Video stream */}
-        {displayMode === 'video' ||
-        (displayMode === 'auto' && useVideoStream && !videoStreamFailed) ? (
-          <ScrcpyPlayer
-            deviceId={deviceId}
-            className="w-full h-full"
-            enableControl={true}
-            onFallback={handleFallback}
-            onTapSuccess={() => showFeedback(t.devicePanel.tapped, 2000, 'tap')}
-            onTapError={error =>
-              showFeedback(
-                t.devicePanel.tapError.replace('{error}', error),
-                3000,
-                'error'
-              )
-            }
-            onSwipeSuccess={() =>
-              showFeedback(t.devicePanel.swiped, 2000, 'swipe')
-            }
-            onSwipeError={error =>
-              showFeedback(
-                t.devicePanel.swipeError.replace('{error}', error),
-                3000,
-                'error'
-              )
-            }
-            onStreamReady={handleVideoStreamReady}
-            fallbackTimeout={10000}
-          />
-        ) : (
-          /* Screenshot mode */
-          <div className="w-full h-full flex items-center justify-center bg-muted/30 min-h-0">
-            {screenshot && screenshot.success ? (
-              <div className="relative w-full h-full flex items-center justify-center min-h-0">
-                <img
-                  src={`data:image/png;base64,${screenshot.image}`}
-                  alt="Device Screenshot"
-                  className="max-w-full max-h-full object-contain"
-                  style={{
-                    width:
-                      screenshot.width > screenshot.height ? '100%' : 'auto',
-                    height:
-                      screenshot.width > screenshot.height ? 'auto' : '100%',
-                  }}
-                />
-                {screenshot.is_sensitive && (
-                  <div className="absolute top-12 right-2 px-2 py-1 bg-yellow-500 text-white text-xs rounded-lg">
-                    {t.devicePanel.sensitiveContent}
-                  </div>
-                )}
-              </div>
-            ) : screenshot?.error ? (
-              <div className="text-center text-destructive">
-                <AlertCircle className="w-8 h-8 mx-auto mb-2" />
-                <p className="font-medium">{t.devicePanel.screenshotFailed}</p>
-                <p className="text-xs mt-1 opacity-60">{screenshot.error}</p>
-              </div>
-            ) : (
-              <div className="text-center text-muted-foreground">
-                <Loader2 className="w-8 h-8 mx-auto mb-2 animate-spin" />
-                <p className="text-sm">{t.devicePanel.loading}</p>
-              </div>
-            )}
-          </div>
-        )}
-      </Card>
+      <DeviceMonitor deviceId={deviceId} isVisible={true} />
     </div>
   );
 }
